backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Abhishek <abhishek@saxecap.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "3.0.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
Flask-SQLAlchemy = "^3.1.1"
Flask-Migrate = "^4.0.7"
pytest = "^8.1.1"
marshmallow = "^3.21.1"
boto3 = "^1.34.108"
pgvector = "^0.2.5"
flask-limiter = {extras = ["redis"], version = "^3.7.0"}
redis = {extras = ["hiredis"], version = "^5.0.4"}
tiktoken = "^0.7.0"
openai = "^1.30.3"
pandas = "^2.2.2"
flask-cors = "^4.0.1"
gunicorn = "^22.0.0"
amqpstorm = "^2.10.7"
sentry-sdk = {extras = ["openai"], version = "^2.6.0"}

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = {version = "^21.7b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
